update unit tests for new config patterns
